UPDAT

+++ sigrc-backend/src/main/java/com/epmapa/sigrc/domain/service/CorrespondenciaService.java
@@ -847,7 +847,7 @@ public class CorrespondenciaService {
                 areaIds,
                 areaNombres,
                 listarAdjuntos(entity.getIdCorrespondencia()),
-                adjuntoRepository.countByCorrespondenciaIdCorrespondencia(entity.getIdCorrespondencia()),
+                (int) adjuntoRepository.countByCorrespondenciaIdCorrespondencia(entity.getIdCorrespondencia()),
                 obtenerHistorial(entity.getIdCorrespondencia()),
                 obtenerRespuestas(entity.getIdCorrespondencia()),
                 obtenerTicketsVinculados(entity.getIdCorrespondencia()),
